# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "M3th",
        "description": "math testing",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://node43.webte.fei.stuba.sk",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/user-tests": {
            "get": {
                "tags": [
                    "User Tests"
                ],
                "summary": "List all attempts for the authenticated user",
                "operationId": "f210ba888ddfdb21b8f17750cb50b0c5",
                "responses": {
                    "200": {
                        "description": "List of test attempts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "history_tests": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/HistoryTest"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Tests"
                ],
                "summary": "Create a new test attempt",
                "operationId": "593a955752b0be6d3f3fc872e7551c1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "test_id"
                                ],
                                "properties": {
                                    "test_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Bratislava"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "Slovakia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attempt created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HistoryTest"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user-tests/{user_test}": {
            "get": {
                "tags": [
                    "User Tests"
                ],
                "summary": "Show one test attempt with its question records",
                "operationId": "d7a1c82332f9683675b4f8c7d5892c92",
                "parameters": [
                    {
                        "name": "user_test",
                        "in": "path",
                        "description": "ID of the user test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User test attempt with question IDs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "test_id": {
                                            "type": "integer"
                                        },
                                        "score": {
                                            "type": "integer"
                                        },
                                        "city": {
                                            "type": "string"
                                        },
                                        "state": {
                                            "type": "string"
                                        },
                                        "question_ids": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Tests"
                ],
                "summary": "Update score or location on a test attempt",
                "operationId": "b64b933a7d49b0964f5c305f05ec9e66",
                "parameters": [
                    {
                        "name": "user_test",
                        "in": "path",
                        "description": "ID of the user test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "score"
                                ],
                                "properties": {
                                    "score": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Bratislava"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "Slovakia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User test updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HistoryTest"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Tests"
                ],
                "summary": "Delete a test attempt and its question records",
                "operationId": "244d97a6749e4a01623cb7ba01aaa34f",
                "parameters": [
                    {
                        "name": "user_test",
                        "in": "path",
                        "description": "ID of the user test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attempt deleted message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Attempt deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user-tests/{user_test}/evaluate": {
            "post": {
                "tags": [
                    "User Tests"
                ],
                "summary": "Evaluate a user test attempt and update score",
                "operationId": "7f31a02a7a14c173748ecbea74afac91",
                "parameters": [
                    {
                        "name": "user_test",
                        "in": "path",
                        "description": "ID of the user test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evaluation summary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_questions": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "correct_answers": {
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "score": {
                                            "type": "integer",
                                            "example": 8
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Question": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "assignment_sk": {
                        "type": "string"
                    },
                    "assignment_en": {
                        "type": "string"
                    },
                    "isMultiChoice": {
                        "type": "boolean"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            },
            "Answer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "question_id": {
                        "type": "integer"
                    },
                    "answer_sk": {
                        "type": "string"
                    },
                    "answer_en": {
                        "type": "string"
                    },
                    "isCorrect": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name_sk": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "HistoryTest": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "test_id": {
                        "type": "integer"
                    },
                    "score": {
                        "type": "integer"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}